requirements.txt
Flask
flask_pymongo
bcrypt
flask_cors
sentence_transformers
faiss-cpu
numpy
transformers
torch
scikit-learn
langchain_google_genai
langchain
google-generativeai
python-dotenv
gunicorn
